New upstream version 10.5.18

+++ extra/wolfssl/wolfssl/wolfcrypt/src/port/kcapi/kcapi_hash.c
@@ -1,6 +1,6 @@
 /* kcapi_hash.c
  *
- * Copyright (C) 2006-2020 wolfSSL Inc.
+ * Copyright (C) 2006-2022 wolfSSL Inc.
  *
  * This file is part of wolfSSL.
  *
@@ -164,13 +164,13 @@ static int KcapiHashFinal(wolfssl_KCAPI_Hash* hash, byte* out, word32 outSz,
         heap = hash->heap; /* keep because KcapiHashInit clears the pointer */
     #ifdef WOLFSSL_KCAPI_HASH_KEEP
         /* keep full message to out at end instead of incremental updates */
-        ret = kcapi_md_update(hash->handle, hash->msg, hash->used);
+        ret = (int)kcapi_md_update(hash->handle, hash->msg, hash->used);
         XFREE(hash->msg, heap, DYNAMIC_TYPE_TMP_BUFFER);
         hash->msg = NULL;
     #endif
 
         if (ret == 0) {
-            ret = kcapi_md_final(hash->handle, out, outSz);
+            ret = (int)kcapi_md_final(hash->handle, out, outSz);
         }
 
         KcapiHashFree(hash);
@@ -198,9 +198,9 @@ static int KcapiHashGet(wolfssl_KCAPI_Hash* hash, byte* out, word32 outSz)
         ret = kcapi_md_init(&hash->handle, hash->type, 0);
     }
     if (ret == 0) {
-        ret = kcapi_md_update(hash->handle, hash->msg, hash->used);
+        ret = (int)kcapi_md_update(hash->handle, hash->msg, hash->used);
         if (ret >= 0) {
-            ret = kcapi_md_final(hash->handle, out, outSz);
+            ret = (int)kcapi_md_final(hash->handle, out, outSz);
             if (ret >= 0) {
                 ret = 0;
             }
@@ -642,4 +642,3 @@ int wc_Sha512_256Copy(wc_Sha512* src, wc_Sha512* dst)
 #endif /* WOLFSSL_SHA512 */
 
 #endif /* WOLFSSL_KCAPI_HASH */
-
